PackageCreateDialog: tsify, reuse dialog code
